feat: add settings page
