+++ src/content/components/AboutView/AboutView.scss
@@ -1,0 +1,5 @@
+.container {
+  max-width: 900px;
+  margin: 0 auto;
+  padding: 20px 40px;
+}
