+++ src/content/components/Router/Router.scss
@@ -104,3 +104,6 @@
 .icon-hourglass {
   background-image: url("../../img/icons/hourglass.svg");
 }
+.icon-settings {
+  background-image: url("../../img/icons/settings.svg");
+}
