+++ src/content/components/SettingsView/SettingsView.scss
